finish project

+++ migrations/20231116222758_create_transactions_table.php
@@ -9,18 +9,18 @@ final class CreateTransactionsTable extends AbstractMigration
     public function change(): void
     {
         $sql = <<< SQL
-CREATE TABLE transaction (
-        id INT PRIMARY KEY AUTO_INCREMENT,
-    payee_id INT NOT NULL,
-    payer_id INT NOT NULL,
-    amount DECIMAL(10, 2) NOT NULL,
-    status VARCHAR(255) NOT NULL,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    finished_at TIMESTAMP DEFAULT NULL,
-    FOREIGN KEY (payee_id) REFERENCES users(id),
-    FOREIGN KEY (payer_id) REFERENCES users(id)
-);
-SQL;
+        CREATE TABLE transaction (
+            id INT PRIMARY KEY AUTO_INCREMENT,
+            payee_id INT NOT NULL,
+            payer_id INT NOT NULL,
+            amount INT NOT NULL,
+            status VARCHAR(255) NOT NULL,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            finished_at TIMESTAMP DEFAULT NULL,
+            FOREIGN KEY (payee_id) REFERENCES users(id),
+            FOREIGN KEY (payer_id) REFERENCES users(id)
+        );
+        SQL;
         $this->execute($sql);
     }
 }
